Yg-129 🔖 Rename : post관련 전역상태파일 이름 수정
