+++ app/(afterLogin)/detailPost/[postId]/_components/floating/type.ts
@@ -14,6 +14,6 @@ export type FloatingButtonType = {
 export type FloatingBarProps = {
     icons: { name: string; icon: string; isActive: boolean }[]
     isMine?: boolean
-    postId?: string
-    post?: Post
+    postId: number 
+    post?: Post | undefined
 }
