+++ hook/type.ts
@@ -29,7 +29,8 @@ export type SigninResult = UseMutationResult<UserResponse, Error, UserRequest> &
     handleOverlay: (isOpen: boolean, state?: "success" | "fail" | null) => void
 }
 export type useHandleClickProps = {
-    postId?: string
+    postId: number
+    liked? : boolean
     post?: Post
     setIconState: Dispatch<SetStateAction<FloatingIcon[]>>
 }
